Fixed hard crash caused by bad API return
The enums for pet gender only accounted for "M" and "F" coming back from the API.  In this specific crash, we were getting back "U".  Because the enum initializer was being force unwrapped, we got a crash.

I added in "unknown" enum cases for our enums and fixed the failable initializers to default to this unknown case if the initializer fails.

Also, I noticed a problem with breeds for some animals coming back as "unknown" on the detail VC.  This was because of a simple typo and has also been fixed.