Infinite Scrolling has been implmented
No longer are we constrained to the first 25 pets returned by PetFinder!  Infinite scrolling has been implemented to allow the app to retrieve the next set of results when they reach the bottom of the collection view.  Due to a usage limit imposed by PetFinder, we are only ever able to see 2000 search results.  That constraint has been added to the code as well.

While implementing these changes, I also cleaned up some code, including creating a Constants file for constants that might be needed throughout, as well as updating the PTKRequestManager to accommodate for retrieving and passing an offset to and from the API, and adjusting the code in the PetListingViewController to handle the infinite scroll in a UX-friendly manner (added an activity indicator to the bottom that will show as new results are being loaded in, etc)